+++ app/Http/Controllers/Api/StudentController.php
@@ -108,11 +108,17 @@ class StudentController extends Controller
         $user->update(["auth_token" => null]);
         return api_response(1, __("site.student signOut successfully"));
     }
-    public function deleteAccount()
+    public function deleteAccount(Request $request)
     {
+        $request->validate([
+            "password" => ["required","confirmed"]
+        ]);
         $user = auth("api")->user();
-        $user->delete();
-        return api_response(1, __("site.student deleted successfully"));
+        if (Hash::check($request->password,$user->password)){
+            $user->delete();
+            return api_response(1, __("site.student deleted successfully"));
+        }
+        return api_response(0, __("site.Sorry Wrong Password"));
     }
     public function getPosts(){
         $posts = Post::active()->withCount("replies")->latest()->paginate(6);
